requirements.txt
spotipy
python-dotenv


